- Update MapContainer to better support draw controls - Update MapPage to pass additional draw control options to MapContainer - Move MapPage spatial constants and util functions to separate files
