+++ app/src/features/map/MapPage.tsx
@@ -2,45 +2,18 @@ import Box from '@material-ui/core/Box';
 import Container from '@material-ui/core/Container';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import { BoundaryFeature, BoundaryFeaturePopup } from 'components/map/BoundaryFeaturePopup';
 import { IMarkerLayer } from 'components/map/components/MarkerCluster';
 import { IStaticLayer } from 'components/map/components/StaticLayers';
 import MapContainer from 'components/map/MapContainer';
-import { OccurrenceFeature, OccurrenceFeaturePopup } from 'components/map/OccurrenceFeaturePopup';
+import { ALL_OF_BC_BOUNDARY, SPATIAL_COMPONENT_TYPE } from 'constants/spatial';
 import { Feature } from 'geojson';
 import { useApi } from 'hooks/useApi';
 import useDataLoader from 'hooks/useDataLoader';
 import useDataLoaderError from 'hooks/useDataLoaderError';
-import { LatLngBounds, LatLngTuple } from 'leaflet';
+import { LatLngBounds } from 'leaflet';
 import React, { useEffect, useState } from 'react';
+import { parseFeatureCollectionsByType } from 'utils/spatial-utils';
 import { getFeatureObjectFromLatLngBounds } from 'utils/Utils';
-
-export const ALL_OF_BC_BOUNDARY: Feature = {
-  type: 'Feature',
-  properties: {},
-  geometry: {
-    type: 'Polygon',
-    coordinates: [
-      [
-        [-146.95401365536304, 44.62175409623327],
-        [-146.95401365536304, 63.528970541102794],
-        [-105.07413084286304, 63.528970541102794],
-        [-105.07413084286304, 44.62175409623327],
-        [-146.95401365536304, 44.62175409623327]
-      ]
-    ]
-  }
-};
-
-export enum SPATIAL_COMPONENT_TYPE {
-  OCCURRENCE = 'Occurrence',
-  BOUNDARY = 'Boundary'
-}
-
-export enum LAYER_NAME {
-  OCCURRENCES = 'Occurrences',
-  BOUNDARIES = 'Boundaries'
-}
 
 const MapPage: React.FC = () => {
   const api = useApi();
@@ -65,42 +38,10 @@ const MapPage: React.FC = () => {
       return;
     }
 
-    const occurrencesMarkerLayer: IMarkerLayer = { layerName: LAYER_NAME.OCCURRENCES, markers: [] };
+    const result = parseFeatureCollectionsByType(mapDataLoader.data);
 
-    const occurrenceStaticLayer: IStaticLayer = { layerName: LAYER_NAME.OCCURRENCES, features: [] };
-    const boundaryStaticLayer: IStaticLayer = { layerName: LAYER_NAME.BOUNDARIES, features: [] };
-
-    for (const featureCollection of mapDataLoader.data) {
-      for (const feature of featureCollection.features) {
-        if (isOccurrenceFeature(feature)) {
-          if (feature.geometry.type === 'GeometryCollection') {
-            // Not expecting or supporting geometry collections
-            continue;
-          }
-
-          occurrencesMarkerLayer.markers.push({
-            position: feature.geometry.coordinates as LatLngTuple,
-            key: feature.id,
-            popup: <OccurrenceFeaturePopup properties={feature.properties} />
-          });
-        }
-
-        if (isBoundaryFeature(feature)) {
-          if (feature.geometry.type === 'GeometryCollection') {
-            continue;
-          }
-
-          boundaryStaticLayer.features.push({
-            geoJSON: feature,
-            key: feature.id,
-            popup: <BoundaryFeaturePopup properties={feature.properties} />
-          });
-        }
-      }
-    }
-
-    setStaticLayers([occurrenceStaticLayer, boundaryStaticLayer]);
-    setMarkerLayers([occurrencesMarkerLayer]);
+    setStaticLayers(result.staticLayers);
+    setMarkerLayers(result.markerLayers);
   }, [mapDataLoader.data]);
 
   mapDataLoader.load(ALL_OF_BC_BOUNDARY, [SPATIAL_COMPONENT_TYPE.BOUNDARY, SPATIAL_COMPONENT_TYPE.OCCURRENCE]);
@@ -124,6 +65,16 @@ const MapPage: React.FC = () => {
                       SPATIAL_COMPONENT_TYPE.OCCURRENCE
                     ]);
                   }}
+                  drawControls={{
+                    options: {
+                      draw: { circle: false, circlemarker: false, marker: false, polyline: false },
+                      edit: { edit: false, remove: false }
+                    },
+                    onChange: (features: Feature[]) => {
+                      console.log('Drawn features', features);
+                    },
+                    clearOnDraw: true
+                  }}
                   scrollWheelZoom={true}
                   markerLayers={markerLayers}
                   staticLayers={staticLayers}
@@ -138,11 +89,3 @@ const MapPage: React.FC = () => {
 };
 
 export default MapPage;
-
-export const isOccurrenceFeature = (feature: Feature): feature is OccurrenceFeature => {
-  return feature.geometry.type === 'Point' && feature.properties?.['type'] === SPATIAL_COMPONENT_TYPE.OCCURRENCE;
-};
-
-export const isBoundaryFeature = (feature: Feature): feature is BoundaryFeature => {
-  return feature?.properties?.['type'] === SPATIAL_COMPONENT_TYPE.BOUNDARY;
-};
